pyproject.toml
[project]
name = "basic-unsupervised-learning-algorithms-and-text-data-processing"
version = "0.1.0"
description = ""
authors = [
    {name = "Alduin",email = "prime72w@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.3.4,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "wordcloud (>=1.9.4,<2.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "mplcyberpunk (>=0.7.6,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
